new requirements

+++ marquespoint_overall/wizard/installment_wizard.py
@@ -1,7 +1,7 @@
-from odoo import fields, models, _, api
+from odoo import fields, models, api, _
 from datetime import date
 from dateutil.relativedelta import relativedelta
-from odoo.exceptions import Warning, UserError
+from odoo.exceptions import Warning, UserError, ValidationError
 
 
 class InstallmentWizard(models.TransientModel):
@@ -35,32 +35,33 @@ class InstallmentWizard(models.TransientModel):
     @api.depends('percentage', 'order_id.amount_untaxed', 'milestone_id.amount', 'is_token_money', 'is_admin_fee',
                  'admin_fee')
     def _compute_amount(self):
-        if self.percentage and self.is_admin_fee:
+        if self.is_admin_fee:
             self.amount = self.admin_fee
+            self.percentage = 0
         elif self.percentage and self.order_id.amount_untaxed:
             self.amount = self.order_id.amount_untaxed * (self.percentage / 100.0)
             if self.is_token_money:
                 if temp := sum(
-                    payment.amount
-                    for payment in self.order_id.account_payment_ids
-                    if payment.state == 'post'
+                        payment.amount
+                        for payment in self.order_id.account_payment_ids
+                        if payment.state == 'post'
                 ):
                     self.amount -= temp
-                    product_template = self.env['product.template'].search(
+                    product_unit = self.env['product.template'].search(
                         [('name', '=', self.order_id.partner_id.name)])
-                    product_template.status = 'reserved'
+                    product_unit.state = 'reserved'
         elif self.percentage and self.milestone_id.amount:
             self.amount = self.milestone_id.amount * (self.percentage / 100.0)
             if self.is_token_money:
                 if temp := sum(
-                    payment.amount
-                    for payment in self.order_id.account_payment_ids
-                    if payment.state == 'post'
+                        payment.amount
+                        for payment in self.order_id.account_payment_ids
+                        if payment.state == 'post'
                 ):
                     self.amount -= temp
-                    product_template = self.env['product.template'].search(
+                    product_unit = self.env['product.product'].search(
                         [('name', '=', self.order_id.partner_id.name)])
-                    product_template.status = 'reserved'
+                    product_unit.state = 'reserved'
         else:
             self.amount = 0
 
@@ -82,6 +83,8 @@ class InstallmentWizard(models.TransientModel):
     def create_installments(self):
         if self.amount == 0 or self.installment_no == 0:
             raise UserError('Amount or Installment No are Missing')
+        if self.percentage < 0:
+            raise ValidationError('Please Enter Positive value for Percentage')
         model = self.env.context.get('active_model')
         active_id = self.env[model].browse(self.env.context.get('active_id'))
         self.env['installment.line'].search([('milestone_id', '=', self.milestone_id.id)]).unlink()
@@ -93,51 +96,35 @@ class InstallmentWizard(models.TransientModel):
             'installment_no': self.installment_no,
             'installment_period': self.installment_period,
         })
-        amount = self.amount / self.installment_no
+        # Validations
+        if active_id.order_id.plan_ids:
+            total_percentage = sum(percen.percentage for percen in active_id.order_id.plan_ids)
+            total_amount = sum(line.amount for line in active_id.order_id.plan_ids)
+            print(f'total_percentage: {total_percentage}')
+            print(f'total_amount: {total_amount}')
+            print(f'self.order_id.amount_untaxed: {active_id.order_id.amount_untaxed}')
+            remaining_percentage = 100 - (total_percentage - self.percentage)
+            if total_amount > active_id.order_id.amount_untaxed:
+                raise ValidationError('The amount is Exceeding the sale order')
+            if total_percentage > 100:
+                raise ValidationError(
+                    f'Capacity: 100%\nFilled: {total_percentage - self.percentage}% \nRemaining: {remaining_percentage}%')
+        #     End validations
+        # amount = self.amount / self.installment_no
         for _ in range(self.installment_no):
             inv_date = self.start_date
-            invoice = False
-            if self.order_id.order_line:
-                inv_lines = [
-                    (
-                        0,
-                        0,
-                        {
-                            'product_id': line.product_id.id,
-                            'name': line.name,
-                            'quantity': line.product_uom_qty,
-                            'product_uom_id': line.product_uom.id,
-                            'price_unit': amount,
-                            'tax_ids': line.tax_id,
-                        },
-                    )
-                    for line in self.order_id.order_line
-                ]
-                if self.installment_period == 'month':
-                    inv_date += relativedelta(months=_)
-                    print(f'inv date: {inv_date}')
-                elif self.installment_period == 'quarterly':
-                    inv_date += relativedelta(months=_*3)
-                elif self.installment_period == 'annual':
-                    inv_date += relativedelta(months=_*12)
-                inv_vals = {
-                    'partner_id': self.order_id.partner_id.id,
-                    'invoice_date': inv_date,
-                    'invoice_line_ids': inv_lines,
-                    'move_type': 'out_invoice',
-                    'so_ids': self.order_id.id,
-                    'state': 'draft',
-                    'project': self.order_id.project.id,
-                    'building': self.order_id.building.id,
-                    'floor': self.order_id.floor.id,
-                    'unit': self.order_id.unit.id,
-                }
-                invoice = self.env['account.move'].create(inv_vals)
+            if self.installment_period == 'month':
+                inv_date += relativedelta(months=_)
+                print(f'inv date: {inv_date}')
+            elif self.installment_period == 'quarterly':
+                inv_date += relativedelta(months=_ * 3)
+            elif self.installment_period == 'annual':
+                inv_date += relativedelta(months=_ * 12)
             vals = {
                 'milestone_id': self.milestone_id.id,
-                'amount': amount,
+                # 'amount': amount,
                 'order_id': self.order_id.id,
-                'move_id': invoice.id
+                'date': inv_date
             }
             self.env['installment.line'].create(vals)
         print('installment created')
